refactor file structure (group handlers with services)
