+++ src/main/java/org/example/entities/connection/Connection.java
@@ -1,4 +1,4 @@
-package org.example.entities;
+package org.example.entities.connection;
 
 import com.google.gson.annotations.Expose;
 import java.util.Objects;
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
+import org.example.entities.DataTransferObject;
 
 @Getter
 @Setter
