+++ src/main/java/org/example/handlers/rest/login/LoginHandler.java
@@ -1,4 +1,4 @@
-package org.example.handlers.rest;
+package org.example.handlers.rest.login;
 
 import static org.example.utils.APIGatewayResponseBuilder.makeHttpResponse;
 
@@ -14,7 +14,6 @@ import org.example.exceptions.Unauthorized;
 import org.example.models.requests.LoginRequest;
 import org.example.models.requests.SessionRequest;
 import org.example.models.responses.rest.LoginResponseBody;
-import org.example.services.LoginService;
 import org.example.utils.ValidateObject;
 
 public class LoginHandler
