+++ src/main/java/org/example/handlers/rest/login/LoginService.java
@@ -1,4 +1,4 @@
-package org.example.services;
+package org.example.handlers.rest.login;
 
 import static com.mongodb.client.model.Filters.eq;
