+++ src/main/java/org/example/handlers/rest/stats/StatsHandler.java
@@ -1,4 +1,4 @@
-package org.example.handlers.rest;
+package org.example.handlers.rest.stats;
 
 import static org.example.utils.APIGatewayResponseBuilder.makeHttpResponse;
