+++ src/main/java/org/example/handlers/websocket/connect/ConnectService.java
@@ -1,4 +1,4 @@
-package org.example.services;
+package org.example.handlers.websocket.connect;
 
 import com.mongodb.client.model.Filters;
 import java.util.List;
