+++ src/main/java/org/example/handlers/websocket/defaultRoute/DefaultHandler.java
@@ -1,4 +1,4 @@
-package org.example.handlers.websocket;
+package org.example.handlers.websocket.defaultRoute;
 
 import static org.example.utils.APIGatewayResponseBuilder.makeWebsocketResponse;
 
@@ -7,29 +7,12 @@ import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayV2WebSocketEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayV2WebSocketResponse;
 import org.example.constants.StatusCodes;
-import org.example.entities.Connection;
-import org.example.utils.MongoDBUtility;
 
-public class DisconnectHandler
+public class DefaultHandler
     implements RequestHandler<APIGatewayV2WebSocketEvent, APIGatewayV2WebSocketResponse> {
-
-  MongoDBUtility<Connection> utility;
-
-  public DisconnectHandler() {
-    this.utility = new MongoDBUtility<>("connections", Connection.class);
-  }
-
-  public DisconnectHandler(MongoDBUtility<Connection> utility) {
-    this.utility = utility;
-  }
-
   @Override
   public APIGatewayV2WebSocketResponse handleRequest(
       APIGatewayV2WebSocketEvent event, Context context) {
-    String connectionId = event.getRequestContext().getConnectionId();
-
-    utility.delete(connectionId);
-
-    return makeWebsocketResponse(StatusCodes.OK, "Successfully disconnected");
+    return makeWebsocketResponse(StatusCodes.BAD_REQUEST, "Route: is not supported");
   }
 }
