+++ src/test/java/org/example/handlers/auth/AuthHandlerTest.java
@@ -8,7 +8,7 @@ import java.util.HashMap;
 import java.util.Map;
 import org.example.entities.session.Session;
 import org.example.entities.user.User;
-import org.example.handlers.rest.AuthHandler;
+import org.example.handlers.rest.auth.AuthHandler;
 import org.example.utils.MockContext;
 import org.example.utils.MongoDBUtility;
 import org.junit.jupiter.api.*;
