+++ src/test/java/org/example/handlers/disconnect/DisconnectHandlerTest.java
@@ -7,8 +7,8 @@ import com.amazonaws.services.lambda.runtime.events.APIGatewayV2WebSocketEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayV2WebSocketResponse;
 import java.util.Optional;
 import org.example.constants.StatusCodes;
-import org.example.entities.Connection;
-import org.example.handlers.websocket.DisconnectHandler;
+import org.example.entities.connection.Connection;
+import org.example.handlers.websocket.disconnect.DisconnectHandler;
 import org.example.utils.MockContext;
 import org.example.utils.MongoDBUtility;
 import org.junit.jupiter.api.BeforeAll;
