+++ src/test/java/org/example/handlers/joinGame/JoinGameHandlerTest.java
@@ -16,9 +16,9 @@ import org.example.entities.user.User;
 import org.example.enums.GameStatus;
 import org.example.enums.TimeControl;
 import org.example.exceptions.NotFound;
-import org.example.handlers.websocket.JoinGameHandler;
+import org.example.handlers.websocket.joinGame.JoinGameHandler;
+import org.example.handlers.websocket.joinGame.JoinGameService;
 import org.example.models.requests.JoinGameRequest;
-import org.example.services.JoinGameService;
 import org.example.utils.MockContext;
 import org.example.utils.MongoDBUtility;
 import org.example.utils.socketMessenger.SocketSystemLogger;
