+++ src/test/java/org/example/handlers/logout/LogoutHandlerTest.java
@@ -19,9 +19,9 @@ import org.example.entities.user.UserDbService;
 import org.example.enums.GameStatus;
 import org.example.enums.TimeControl;
 import org.example.exceptions.NotFound;
-import org.example.handlers.rest.LogoutHandler;
-import org.example.services.GameStateService;
-import org.example.services.LogoutService;
+import org.example.handlers.rest.getGameState.GameStateService;
+import org.example.handlers.rest.logout.LogoutHandler;
+import org.example.handlers.rest.logout.LogoutService;
 import org.example.utils.MockContext;
 import org.example.utils.socketMessenger.SocketSystemLogger;
 import org.junit.jupiter.api.*;
