+++ src/test/java/org/example/handlers/makeMove/MakeMoveHandlerTest.java
@@ -17,18 +17,18 @@ import org.example.entities.move.Move;
 import org.example.entities.player.Player;
 import org.example.entities.stats.Stats;
 import org.example.entities.user.User;
-import org.example.enums.WebsocketResponseAction;
 import org.example.enums.TimeControl;
+import org.example.enums.WebsocketResponseAction;
 import org.example.exceptions.NotFound;
-import org.example.handlers.websocket.JoinGameHandler;
-import org.example.handlers.websocket.MakeMoveHandler;
+import org.example.handlers.websocket.joinGame.JoinGameHandler;
+import org.example.handlers.websocket.joinGame.JoinGameService;
+import org.example.handlers.websocket.makeMove.MakeMoveHandler;
+import org.example.handlers.websocket.makeMove.MakeMoveService;
 import org.example.models.requests.JoinGameRequest;
 import org.example.models.requests.MakeMoveRequest;
 import org.example.models.responses.websocket.GameStartedMessageData;
 import org.example.models.responses.websocket.MakeMoveMessageData;
 import org.example.models.responses.websocket.SocketResponseBody;
-import org.example.services.JoinGameService;
-import org.example.services.MakeMoveService;
 import org.example.utils.MockContext;
 import org.example.utils.MongoDBUtility;
 import org.example.utils.socketMessenger.SocketSystemLogger;
