+++ src/test/java/org/example/handlers/register/RegisterHandlerTest.java
@@ -13,9 +13,9 @@ import org.bson.types.ObjectId;
 import org.example.constants.StatusCodes;
 import org.example.entities.stats.Stats;
 import org.example.entities.user.User;
-import org.example.handlers.rest.RegisterHandler;
+import org.example.handlers.rest.register.RegisterHandler;
+import org.example.handlers.rest.register.RegisterService;
 import org.example.models.requests.RegisterRequest;
-import org.example.services.RegisterService;
 import org.example.utils.EncryptPassword;
 import org.example.utils.MockContext;
 import org.example.utils.MongoDBUtility;
