+++ src/test/java/org/example/utils/IntegrationTestUtils.java
@@ -37,18 +37,14 @@ public final class IntegrationTestUtils<T> {
         .asString();
   }
 
-  public Response post(
-          String endpoint,
-          Map<String, String> headers,
-          int statusCode
-  ) {
+  public Response post(String endpoint, Map<String, String> headers, int statusCode) {
     return given()
-            .headers(headers)
-            .post(endpoint)
-            .then()
-            .statusCode(statusCode)
-            .extract()
-            .response();
+        .headers(headers)
+        .post(endpoint)
+        .then()
+        .statusCode(statusCode)
+        .extract()
+        .response();
   }
 
   public Response get(
